pyproject.toml
[project]
name = "qestify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "mysql-connector>=2.2.9",
    "mysql-connector-python>=9.4.0",
    "pymysql>=1.1.2",
]
